+++ test/testdata/demo_long.py
@@ -201,8 +201,6 @@ CONST_NO_DOC = "SHOULD NOT APPEAR"
 class DataDemo:
     """
     This is an example for a dataclass.
-    Dataclasses generate a relatively pointless docstring by default,
-    but you can override it by providing your own (like here!).
 
     As usual, you can link to individual properties: `DataDemo.a`.
     """
@@ -210,9 +208,19 @@ class DataDemo:
     a: int
     """Again, we can document individual properties with docstrings."""
     a2: Sequence[str]
-    # This property has a type annotation but is not documented, so it does not show up.
+    # This property has a type annotation but is not documented.
+    a3 = "a3"
+    # This property has a default value but is not documented.
+    a4: str = "a4"
+    # This property has a type annotation and a default value but is not documented.
     b: bool = field(repr=False, default=True)
     """This property is assigned to `dataclasses.field()`, which works just as well."""
+
+
+@dataclass
+class DataDemoExtended(DataDemo):
+    c: str = "42"
+    """A new attribute."""
 
 
 class EnumDemo(enum.Enum):