+++ test/testdata/demopackage/child_c.py
@@ -4,5 +4,6 @@ from . import child_b
 
 class C(child_b.B):
     """This class is defined in .child_c and inherits from .child_b.B"""
+
     def c(self):
         return 2