+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 # Unreleased: pdoc next
 
+ - Make documentation of variables more consistent. Variables with a default value
+   and no docstring are now hidden, matching the behavior of variables with a type annotation only.
+   ([#411](https://github.com/mitmproxy/pdoc/issues/411), [@mhils](https://github.com/mhils))
  - Remove `format` argument from `pdoc.pdoc()`. For the forseeable future, pdoc will only support HTML export.
    ([#308](https://github.com/mitmproxy/pdoc/issues/308), [@mhils](https://github.com/mhils))
  - Update vendored copy of markdown2.
    ([#429](https://github.com/mitmproxy/pdoc/issues/429), [@mhils](https://github.com/mhils))
  - Fix "View Source" button when a function has the same name as the module it is in.
    ([#431](https://github.com/mitmproxy/pdoc/issues/431), [@mhils](https://github.com/mhils))
+ - Improve display of dataclasses.
+   ([#411](https://github.com/mitmproxy/pdoc/issues/411), [@mhils](https://github.com/mhils))
 
 # 2022-06-08: pdoc 12.1.0